touchscreen: goodix_gt9xx: fix coverity warning
PD#172713: touchscreen: goodix_gt9xx: fix coverity warning

1.The return value of function "gtp_i2c_read" is not
checked in the "gt9xx.c" file.
2."ts"null pointer dereferencing reference in the "gt9xx.c" file.
3.In the "goodix_tool.c" file, because the third argument to
"memset" function is of type int, expression "cmd_head.data_len + 1"
is of type u16, so it need to convert to int.
This causes "Unused value".

+++ drivers/amlogic/input/touchscreen/goodix_gt9xx/goodix_tool.c
@@ -487,7 +487,7 @@ ssize_t goodix_tool_write(struct file *filp, const char __user *buff, size_t len
     {
         show_len = 0;
         total_len = 0;
-        memset(cmd_head.data, 0, cmd_head.data_len + 1);
+        memset(cmd_head.data, 0, (int)(cmd_head.data_len + 1));
         memcpy(cmd_head.data, &buff[CMD_HEAD_LENGTH], cmd_head.data_len);
 
         if (FAIL == gup_update_proc((void*)cmd_head.data))
